move top-level template codegen to HsTopLevel
