+++ fficxx/src/FFICXX/Generate/Code/HsTemplate.hs
@@ -1,20 +1,27 @@
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE RecordWildCards #-}
 
-module FFICXX.Generate.Code.HsTemplate where
+module FFICXX.Generate.Code.HsTemplate
+  ( genTemplateMemberFunctions,
+    genTMFExp,
+    genTMFInstance,
+    genImportInTemplate,
+    genTmplInterface,
+    genImportInTH,
+    genTmplImplementation,
+    genTmplInstance,
+  )
+where
 
 import qualified Data.List as L (foldr1)
 import FFICXX.Generate.Code.Cpp
-  ( genTLTmplFunCpp,
-    genTmplClassCpp,
+  ( genTmplClassCpp,
     genTmplFunCpp,
     genTmplVarCpp,
   )
 import FFICXX.Generate.Code.HsCast (castBody_)
 import FFICXX.Generate.Code.Primitive
-  ( convertCpp2HS,
-    convertCpp2HS4Tmpl,
-    functionSignatureT,
+  ( functionSignatureT,
     functionSignatureTMF,
     functionSignatureTT,
     tmplAccessorToTFun,
@@ -35,7 +42,6 @@ import FFICXX.Generate.Type.Class
   ( Accessor (Getter, Setter),
     Arg (..),
     Class (..),
-    TLTemplate (..),
     TemplateClass (..),
     TemplateFunction (..),
     TemplateMemberFunction (..),
@@ -46,9 +52,7 @@ import FFICXX.Generate.Type.Module
   ( ClassImportHeader (..),
     TemplateClassImportHeader (..),
     TemplateClassSubmoduleType (..),
-    TopLevelImportHeader (..),
   )
-import FFICXX.Generate.Util (firstUpper)
 import FFICXX.Generate.Util.HaskellSrcExts
   ( bracketExp,
     clsDecl,
@@ -73,12 +77,10 @@ import FFICXX.Generate.Util.HaskellSrcExts
     mkTVar,
     mkVar,
     op,
-    parenSplice,
     pbind_,
     qualConDecl,
     qualifier,
     tyPtr,
-    tySplice,
     tyTupleBoxed,
     tyapp,
     tycon,
@@ -496,212 +498,5 @@ genTmplInstance tcih =
               `app` foldl1
                 (\f x -> con "AppT" `app` f `app` x)
                 (v "con" `app` strE (typeclassNameT t) : map v tvars)
-              `app` (v "lst")
-          ]
-
----------------
--- top-level --
----------------
-
-genTLTemplateInterface :: TLTemplate -> [Decl ()]
-genTLTemplateInterface t =
-  [ mkClass cxEmpty (firstUpper (topleveltfunc_name t)) (map mkTBind tps) methods
-  ]
-  where
-    tps = topleveltfunc_params t
-    ctyp = convertCpp2HS Nothing (topleveltfunc_ret t)
-    lst = map (convertCpp2HS Nothing . arg_type) (topleveltfunc_args t)
-    sigdecl = mkFunSig (topleveltfunc_name t) $ foldr1 tyfun (lst <> [tyapp (tycon "IO") ctyp])
-    methods = [clsDecl sigdecl]
-
-genTLTemplateImplementation :: TLTemplate -> [Decl ()]
-genTLTemplateImplementation t =
-  mkFun nh sig (tvars_p ++ [p "suffix"]) rhs (Just bstmts)
-  where
-    v = mkVar
-    p = mkPVar
-    itps = zip ([1 ..] :: [Int]) (topleveltfunc_params t)
-    tvars = map (\(i, _) -> "typ" ++ show i) itps
-    nparams = length itps
-    tparams = if nparams == 1 then tycon "Type" else tyTupleBoxed (replicate nparams (tycon "Type"))
-    sig = foldr1 tyfun [tparams, tycon "String", tyapp (tycon "Q") (tycon "Exp")]
-    tvars_p = if nparams == 1 then map p tvars else [pTuple (map p tvars)]
-    prefix = "TL"
-    nh = "t_" <> topleveltfunc_name t
-    nc = topleveltfunc_name t
-    lit' = strE (prefix <> "_" <> nc)
-    lam = lamE [p "n"] (lit' `app` v "<>" `app` v "n")
-    rhs =
-      app (v "mkTFunc") $
-        let typs = if nparams == 1 then map v tvars else [tuple (map v tvars)]
-         in tuple (typs ++ [v "suffix", lam, v "tyf"])
-    sig' =
-      let e = error "genTLTemplateImplementation"
-          spls = map (tySplice . parenSplice . mkVar) $ topleveltfunc_params t
-          ctyp = convertCpp2HS4Tmpl e Nothing spls (topleveltfunc_ret t)
-          lst = map (convertCpp2HS4Tmpl e Nothing spls . arg_type) (topleveltfunc_args t)
-       in foldr1 tyfun (lst <> [tyapp (tycon "IO") ctyp])
-    tassgns = map (\(i, tp) -> pbind_ (p tp) (v "pure" `app` (v ("typ" ++ show i)))) itps
-    bstmts =
-      binds
-        [ mkBind1
-            "tyf"
-            [wildcard]
-            ( letE
-                tassgns
-                (bracketExp (typeBracket sig'))
-            )
-            Nothing
-        ]
-
-genTLTemplateInstance ::
-  TopLevelImportHeader ->
-  TLTemplate ->
-  [Decl ()]
-genTLTemplateInstance tih t =
-  mkFun
-    fname
-    sig
-    (p "isCprim" : zipWith (\x y -> pTuple [p x, p y]) qtvars pvars)
-    rhs
-    Nothing
-  where
-    p = mkPVar
-    v = mkVar
-    tcname = firstUpper (topleveltfunc_name t)
-    fname = "gen" <> tcname <> "InstanceFor"
-    itps = zip ([1 ..] :: [Int]) (topleveltfunc_params t)
-    tvars = map (\(i, _) -> "typ" ++ show i) itps
-    qtvars = map (\(i, _) -> "qtyp" ++ show i) itps
-    pvars = map (\(i, _) -> "param" ++ show i) itps
-    nparams = length itps
-    typs_v = if nparams == 1 then v (tvars !! 0) else tuple (map v tvars)
-    params_l = listE (map v pvars)
-    sig =
-      foldr1 tyfun $
-        [tycon "IsCPrimitive"]
-          ++ replicate
-            nparams
-            (tyTupleBoxed [tycon "Q" `tyapp` tycon "Type", tycon "TemplateParamInfo"])
-          ++ [tycon "Q" `tyapp` tylist (tycon "Dec")]
-    -- nvfs = zip ([1..] :: [Int]) vfs
-
-    --------------------------
-    -- final RHS expression --
-    --------------------------
-    rhs =
-      doE
-        ( [paramsstmt, suffixstmt]
-            <> [ generator (p "callmod_") (v "fmap" `app` v "loc_module" `app` (v "location")),
-                 letStmt
-                   [ pbind_
-                       (p "callmod")
-                       (v "dot2_" `app` v "callmod_")
-                   ]
-               ]
-            <> map genqtypstmt (zip tvars qtvars)
-            <> [genstmt "f" (1 :: Int)]
-            <> [ foreignSrcStmt,
-                 letStmt lststmt,
-                 qualStmt retstmt
-               ]
-        )
-    --------------------------
-    paramsstmt =
-      letStmt
-        [ pbind_
-            (p "params")
-            (v "map" `app` (v "tpinfoSuffix") `app` params_l)
-        ]
-    suffixstmt =
-      letStmt
-        [ pbind_
-            (p "suffix")
-            ( v "concatMap"
-                `app` (lamE [p "x"] (inapp (strE "_") (op "++") (v "tpinfoSuffix" `app` v "x")))
-                `app` params_l
-            )
-        ]
-    genqtypstmt (tvar, qtvar) = generator (p tvar) (v qtvar)
-    genstmt prefix n =
-      generator
-        (p (prefix <> show n))
-        ( v "mkFunc"
-            `app` strE (topleveltfunc_name t)
-            `app` v ("t_" <> topleveltfunc_name t)
-            `app` typs_v
-            `app` v "suffix"
-        )
-    lststmt = [pbind_ (p "lst") (listE [v "f1"])]
-    -- TODO: refactor out the following code.
-    foreignSrcStmt =
-      qualifier $
-        (v "addModFinalizer")
-          `app` ( v "addForeignSource"
-                    `app` con "LangCxx"
-                    `app` ( L.foldr1
-                              (\x y -> inapp x (op "++") y)
-                              [ includeStatic,
-                                {-                        , includeDynamic
-                                                        , namespaceStr -}
-                                strE (tcname <> "_instance"),
-                                paren $
-                                  caseE
-                                    (v "isCprim")
-                                    [ match (p "CPrim") (strE "_s"),
-                                      match (p "NonCPrim") (strE "")
-                                    ],
-                                strE "(",
-                                v "intercalate"
-                                  `app` strE ", "
-                                  `app` paren (inapp (v "callmod") (op ":") (v "params")),
-                                strE ")\n"
-                              ]
-                          )
-                )
-      where
-        -- temporary
-        includeStatic =
-          strE $
-            concatMap
-              (<> "\n")
-              ( [R.renderCMacro (R.Include (HdrName "MacroPatternMatch.h"))]
-                  ++ map
-                    R.renderCMacro
-                    ( map R.Include (tihExtraHeadersInCPP tih)
-                        ++ [genTLTmplFunCpp CPrim t, genTLTmplFunCpp NonCPrim t]
-                    )
-              )
-    {-
-    cxxHeaders = v "concatMap" `app` (v "tpinfoCxxHeaders") `app` params_l
-    cxxNamespaces = v "concatMap" `app` (v "tpinfoCxxNamespaces") `app` params_l
-
-    includeDynamic =
-      letE
-        [ pbind_ (p "headers") cxxHeaders,
-          pbind_
-            (pApp (name "f") [p "x"])
-            (v "renderCMacro" `app` (con "Include" `app` v "x"))
-        ]
-        (v "concatMap" `app` v "f" `app` v "headers")
-
-    namespaceStr =
-      letE
-        [ pbind_ (p "nss") cxxNamespaces,
-          pbind_
-            (pApp (name "f") [p "x"])
-            (v "renderCStmt" `app` (con "UsingNamespace" `app` v "x"))
-        ]
-        (v "concatMap" `app` v "f" `app` v "nss")
-    -}
-    retstmt =
-      v "pure"
-        `app` listE
-          [ v "mkInstance"
-              `app` listE []
-              -- `app` (v "con" `app` strE tcname)
-              `app` foldl1
-                (\f x -> con "AppT" `app` f `app` x)
-                (v "con" `app` strE tcname : map v tvars)
               `app` (v "lst")
           ]
